@@ -6,6 +6,7 @@
   "author": "laosb <github@of.sb>",
   "license": "MIT",
   "private": false,
+  "type": "module",
   "scripts": {
     "build:tree-sitter": "./build_scripts/build_parser.sh /Applications/Nova.app",
     "build:js": "rollup -c && cp README.md LICENSE CHANGELOG.md SvelteNova.novaextension && rm -rf SvelteNova.novaextension/node_modules",
@@ -13,14 +14,14 @@
     "sign:tree-sitter": "./build_scripts/sign_lib.sh"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@rollup/plugin-typescript": "^8.2.0",
-    "@types/nova-editor-node": "^4.1.0",
-    "prettier": "^2.2.1",
-    "rollup": "^2.40.0",
-    "tslib": "^2.1.0",
-    "typescript": "^4.2.2"
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/nova-editor-node": "^5.1.4",
+    "prettier": "^3.2.5",
+    "rollup": "^4.9.6",
+    "tslib": "^2.6.2",
+    "typescript": "^5.3.3"
   },
   "dependencies": {
     "nova-extension-utils": "^1.4.0"
