@@ -2,7 +2,7 @@ import commonjs from '@rollup/plugin-commonjs'
 import { nodeResolve } from '@rollup/plugin-node-resolve'
 import typescript from '@rollup/plugin-typescript'
 
-module.exports = {
+export default {
   input: 'src/main.ts',
 
   plugins: [typescript(), commonjs(), nodeResolve()],
